HYD-4700 Update req to use http or https
Req needs to be more flexible. Currently it's hardcoded to handle https
requests. It should handle both http and https.

We can update req to act more like a factory. Instead of a
global agent per module, we can have a factory that takes in
either http or https. An agent can also be passed in. If not
an agent will be generated. Each agent will be tied to a
specific instance.

  - Update copyrights to 2015.
  - Update req to take an optional transport
    string and agent instance.
  - Update req to use fp and obj.
  - Remove lodash.

Signed-off-by: Joe Grund <joe.grund@intel.com>