HYD-5310 req: set options.rejectUnauthorized = false during request not agent
Due to: https://github.com/nodejs/node-v0.x-archive/pull/7792 setting
rejectUnauthorized: false for the agent will create a different hash key
when sockets are stored on a custom agent. When this happens, We are no
longer able to wait for all sockets being removed, as we will continue
to have a hash key that is inaccessible. Moving rejectUnauthorized =
false; onto the options when creating a request fixes this issue.

  - Move rejectUnauthorized prop into buildOptions.
  - Update tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>