HYD-5483 Req: content-length being sent with transfer-encoding: chunked
When req sends a request it passes content-length and transfer-encoding:
chunked. The HTTP 1.1 spec states that you MUST not have both headers.

Latest 0.10 series node caught this in the HTTP parser and throws an
error.

Change the encoding to transfer-encoding: identity. While we're doing
this, default agent connections to keep-alive. This should improve
performance.

  - Default agent to keepAlive: true.
  - Update request header to transfer-encoding: identity.

Signed-off-by: Joe Grund <joe.grund@intel.com>