HYD-5665 Req: extract buffer-request to handle non-JSON
req currently provides a buffered request. However,
it also does lots of other things unrelated to buffering,
such as only working with JSON responses.

As a first step, extract out the buffering from the JSON
related things. In the future we should break down more of the
extra responsibility in bufferRequest.

  - Switch from jshint to eslint.
  - Remove travis.yml.
  - Run against node 4 instead of 0.12.
  - Switch to intel-jasmine-junit-reporter.
  - Extract out buffer-request.
  - Create buffer-json-request for old functionality.
  - Add integration tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>